% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_faceted_hist_plot.R
\name{create_faceted_hist_plot}
\alias{create_faceted_hist_plot}
\title{Create a faceted plot for a target variable}
\usage{
create_faceted_hist_plot(df, feature)
}
\arguments{
\item{df}{Tidy data containing a list of genres and acoustic features, as a \code{df}.}

\item{feature}{Specified acoustic feature of which we want a visualization, as a \code{character}.}
}
\value{
A faceted plot of 6 histograms, one of each genre.
}
\description{
Creates a faceted plot for the target variable "playlist_genre"
for classification using the Spotify Songs dataset
including any selected features,
combined using the face_grid() function
}
\examples{
x <- RCurl::getURL("https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2020/2020-01-21/spotify_songs.csv")
test_data <- read.csv(text = x)
create_faceted_hist_plot(test_data, 'danceability')
}
