% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classy_read.R
\name{classy_read}
\alias{classy_read}
\title{author: Olivia Pang
date: 2023-03-11
Make tidied dataframe}
\usage{
classy_read(path, col = NULL, ...)
}
\arguments{
\item{path}{A file path to a .csv file}

\item{col}{A column in the data frame to convert to the type Factor}

\item{...}{Columns to select from the data frame}
}
\value{
A data frame with the tidied data.
}
\description{
Reads in a CSV file as a data frame,
converts one specified column to the type factor,
selects specified columns to keep,
checks data for NaNs,
then returns the data frame object for the purpose of Classification.
}
\examples{
classy_read("https://gist.githubusercontent.com/curran/a08a1080b88344b0c8a7/raw/0e7a9b0a5d22642a06d3d5b9bcbad9890c8ee534/iris.csv", "petal_width")
}
